fix config
