+++ server/server.go
@@ -1,47 +1,22 @@
 package server
 
 import (
-	"fmt"
-
 	"github.com/BurntSushi/toml"
+	"github.com/devhoodit/sse-chat/auth"
 	githubAuth "github.com/devhoodit/sse-chat/auth/github"
 	"github.com/devhoodit/sse-chat/database"
 	"github.com/gin-gonic/gin"
+	"github.com/redis/go-redis/v9"
 	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/github"
 )
 
-type conf struct {
-	Database *databaseInfo
-	Sentry   *sentryInfo
-	Secret   *secret
-	Oauth    map[string]oauth2Info
-}
-
-type databaseInfo struct {
-	Username string `toml:"username"`
-	Password string `toml:"password"`
-	Address  string `toml:"address"`
-	Port     string `toml:"port"`
-}
-
-type sentryInfo struct {
-	Dsn string
-}
-
-type secret struct {
-	Session string
-}
-
-type oauth2Info struct {
-	Id       string
-	Secret   string
-	Redirect string
-}
+var SecretKey *secret
 
 func SetupRouter() *gin.Engine {
 	r := gin.Default()
 
+	// load config form conf.toml
 	var conf conf
 	_, err := toml.DecodeFile("conf.toml", &conf)
 	if err != nil {
@@ -61,12 +36,22 @@ func SetupRouter() *gin.Engine {
 	// MiddleWare setting, server/middleware.go
 	setMiddleWare(r, &conf)
 
-	// make router
+	rdb := redis.NewClient(&redis.Options{
+		Addr:     "localhost:6379",
+		Password: "", // no password set
+		DB:       0,  // use default DB
+	})
+
+	auth := &auth.JwtAuth{
+		Secret:  conf.Secret.Jwtkey,
+		DB:      d,
+		Session: rdb,
+	}
 
 	// github Oauth router
-	fmt.Println(conf.Oauth)
 	githubRouter := githubAuth.Github{
-		DB: d,
+		DB:      d,
+		JwtAuth: auth,
 		OAuthConfig: &oauth2.Config{
 			ClientID:     conf.Oauth["github"].Id,
 			ClientSecret: conf.Oauth["github"].Secret,
@@ -78,18 +63,19 @@ func SetupRouter() *gin.Engine {
 	r.GET("/", func(c *gin.Context) {
 	})
 
-	auth := r.Group("/auth")
+	authRouter := r.Group("/auth")
 	{
-		sse := auth.Group("/sse")
+		sse := authRouter.Group("/sse")
 		{
-			sse.POST("/login")
+			sse.GET("/login")
 			sse.POST("/create")
 		}
-		github := auth.Group("/github")
+		github := authRouter.Group("/github")
 		{
 			github.GET("/login", githubRouter.Login)
 			github.GET("/callback", githubRouter.Callback)
 		}
+		r.GET("/token").Use().Use()
 	}
 
 	return r
